up net

+++ start.sh
@@ -1,3 +1,7 @@
 #!/bin/sh
-killall lemonbar
-python ~/lemonbar/__init__.py
+PIDS=$(ps -ef | rg "lemonbar -" | rg -v rg | awk '{print $2}')
+if [ ${#PIDS} == 0 ]; then
+  python ~/lemonbar/__init__.py
+else
+  killall lemonbar
+fi
